made it in Java, there are some errors in PHP

+++ Array51/calendar/database_functions.php
@@ -27,21 +27,25 @@ global $connection;
 $query = 'Select id,title, details, eventdate,addedevent from events where eventdate = "'.$day.'/'.$month.'/'.$year.'" order by eventdate';
 			$result = mysqli_query($connection, $query);
 			echo "<table border='1'>";
-			foreach ($result as $value)
+			if($result)
 			{
-				echo "<tr>
-							<td>".$value['id']."</td>
-							<td>".$value['title']."</td>
-							<td>".$value['details']."</td>
-							<td>".$value['eventdate']."</td>
-							<td>".$value['addedevent']."</td>
-							<td><a href='".$_SERVER['PHP_SELF']."?edit=true&id=".$value['id']."'>
-								<input style='height:50px' type ='button' value='EDIT' name ='edit'>
-								</td>
-							<td><a href='".$_SERVER['PHP_SELF']."?delete=true&id=".$value['id']."'>
-								<input style='height:50px' type ='button' value='DELETE' name ='delete'>
-								</td>	
-				     </tr>";
+				foreach ($result as $value)
+				{
+					echo "<tr>
+								<td>".$value['id']."</td>
+								<td>".$value['title']."</td>
+								<td>".$value['details']."</td>
+								<td>".$value['eventdate']."</td>
+								<td>".$value['addedevent']."</td>
+								<td><a href='".$_SERVER['PHP_SELF']."?edit=true&id=".$value['id']."'>
+									<input style='height:50px' type ='button' value='EDIT' name ='edit'>
+									</td>
+								<td><a href='".$_SERVER['PHP_SELF']."?delete=true&id=".$value['id']."'>
+									<input style='height:50px' type ='button' value='DELETE' name ='delete'>
+									</td>	
+						 </tr>";
+				}
+				echo "</table>";
 			}
 			echo "</table>";
 }
@@ -69,7 +73,7 @@ function edit_event_form($day,$month,$year)
 				$event_details = $value['details'];							
 			}
 	
-	echo '<form name="eventform" method = "post" action ='.$_SERVER['PHP_SELF'].'?&edit=true&id='.$id.'>
+	echo '<form name="eventform" method = "post" action ='.$_SERVER['PHP_SELF'].'?edit=true&id='.$id.'>
 		<table width="400px">
 			<tr>
 				<td width="50px">Title</td>
@@ -97,7 +101,7 @@ function update_event($id)
 	$result_update = mysqli_query($connection,$update);
 	if ($result_update)
 	{
-		echo "Event Updated Successfully";
+		echo "Event Updated Successfully. Please Refresh";
 	}
 	else
 	{
